+++ exp/views/__init__.py
@@ -1,6 +1,8 @@
 from exp.views.analytics import *  # noqa
+from exp.views.attachments import *  # noqa
 from exp.views.contact import *  # noqa
 from exp.views.dashboard import *  # noqa
+from exp.views.demographics import *  # noqa
 from exp.views.lab import *  # noqa
 from exp.views.responses import *  # noqa
 from exp.views.study import *  # noqa
